LibWeb/CSS: Stop invalidating style when setting an animation's effect
This reverts f5f71407792f5ff60975c2abf03f6f1ec9a36fc3

Invalidating the style here means that transitions can cause an element
to leave style computation with its "needs style update" flag set to
true. This then causes a VERIFY to fail in the TreeBuilder.

This invalidation does not otherwise seem to have any effect. The
original commit suggests this was to fix a bug, but it's not clear what
bug that was. If it reappears, we can try to solve the issue in a
different way.